refactor(theoremqa): inline the evaluator into the task module
`_theoremqa_eval.py` had exactly one consumer, so it failed the project's
own extraction rule — "do not abstract ahead of time; extract on coupling,
not on call count". The precedent cited for it does not hold either:
`sieval/tasks/_math_verify.py` is imported by 12 task modules, which is the
coupling case the rule describes.

The name was also off-convention. `check_tasks` walks registered task
modules, so `_theoremqa_eval.py` passed by being skipped rather than by
conforming, and it put a second file under a `theoremqa` grep where the
convention is that a task's name is its filename.

No future second consumer was possible: the other known site needing the
same treatment, `sieval/community/deepseek_math.py`, sits in another package
and the private-module rule forbids it reaching into `sieval/tasks/_*`.

Behaviour is unchanged — same 800-sample replay reproduces the stored run
exactly (44.625, empty 2), and the tests move into the task's own test
module, where three cases were added asserting the payload is refused
through each of the three grading entry points rather than only through
`safe_eval`.

Refs #77
