+++ sieval/meta/index.json
@@ -1952,7 +1952,7 @@
       "reference_impl": {
         "source": "TIGER-AI-Lab/TheoremQA",
         "url": "https://github.com/TIGER-AI-Lab/TheoremQA/blob/acfc9686aa9b49f3c8f189364a9a9ee9c53da039/run.py",
-        "notes": "Prompt follows official short-form examples by default; n_shot can select any prefix of the built-in examples. answer_clean and numeric matching follow official utils.py/number_utils.py, with ONE deliberate divergence, taken for execution safety rather than as a repair. UPSTREAM EXECUTES MODEL OUTPUT: number_utils.py turns an extracted answer into a number with a bare eval(num), which is arbitrary code execution driven by the model under test. This task evaluates the same expressions with an AST walk over an allowlist (sieval/tasks/_theoremqa_eval.py) that executes nothing and is bounded in node count, integer-power size and factorial argument. MEASURED DIVERGENCE: ZERO. Replaying a stored 800-sample run (Qwen2.5-72B, 5-shot, greedy) through upstream's bare eval and through this walk reaches 706 expressions across the three call sites; 586 evaluate to an identical value, 120 fail under both, and 0 are refused here but evaluated by upstream. Task accuracy is 44.625 under both readings, and no sample's verdict differs. None of the three bounds is reached by any of the 706. Fidelity was preserved wherever safety did not require otherwise: set displays and bitwise operators are supported purely because upstream computes them, and the builtins that upstream's eval exposed (abs/round/pow/min/max/sum/int/float/complex/divmod/len) are available again -- an earlier mitigation, eval with __builtins__ cleared, was BOTH escapable (clearing builtins does not stop attribute traversal; the catch_warnings route reaches a live open) AND silently lost those names to a NameError. REPRODUCED DEFECT: latex2sympy folds a list answer such as [0, 0] into the set {0}, and evaluating that set cements the corruption; refusing set displays instead would score +2 of 800 samples correct (44.625 -> 44.875) with no verdict moving right-to-wrong. That is a grader repair, not a safety measure, so it is NOT taken here and the _fixed name is left uncoined for a delta this size. GRADING: extraction and comparison both reach latex2sympy, which is synchronous and unbounded, so both run in a worker process (sieval/core/utils/offload.py) rather than on the event loop the session shares."
+        "notes": "Prompt follows official short-form examples by default; n_shot can select any prefix of the built-in examples. answer_clean and numeric matching follow official utils.py/number_utils.py, with ONE deliberate divergence, taken for execution safety rather than as a repair. UPSTREAM EXECUTES MODEL OUTPUT: number_utils.py turns an extracted answer into a number with a bare eval(num), which is arbitrary code execution driven by the model under test. This task evaluates the same expressions with an AST walk over an allowlist (safe_eval, in this module) that executes nothing and is bounded in node count, integer-power size and factorial argument. MEASURED DIVERGENCE: ZERO. Replaying a stored 800-sample run (Qwen2.5-72B, 5-shot, greedy) through upstream's bare eval and through this walk reaches 706 expressions across the three call sites; 586 evaluate to an identical value, 120 fail under both, and 0 are refused here but evaluated by upstream. Task accuracy is 44.625 under both readings, and no sample's verdict differs. None of the three bounds is reached by any of the 706. Fidelity was preserved wherever safety did not require otherwise: set displays and bitwise operators are supported purely because upstream computes them, and the builtins that upstream's eval exposed (abs/round/pow/min/max/sum/int/float/complex/divmod/len) are available again -- an earlier mitigation, eval with __builtins__ cleared, was BOTH escapable (clearing builtins does not stop attribute traversal; the catch_warnings route reaches a live open) AND silently lost those names to a NameError. REPRODUCED DEFECT: latex2sympy folds a list answer such as [0, 0] into the set {0}, and evaluating that set cements the corruption; refusing set displays instead would score +2 of 800 samples correct (44.625 -> 44.875) with no verdict moving right-to-wrong. That is a grader repair, not a safety measure, so it is NOT taken here and the _fixed name is left uncoined for a delta this size. GRADING: extraction and comparison both reach latex2sympy, which is synchronous and unbounded, so both run in a worker process (sieval/core/utils/offload.py) rather than on the event loop the session shares."
       },
       "status": "stable"
     },
