+++ tests/unit/tasks/test_theoremqa_kshot_base_gen.py
@@ -4,9 +4,12 @@ Unit tests for TheoremQA k-shot base generative task.
 AI-Generated Code - GPT-5.5 (OpenAI)
 """
 
+import builtins
 import importlib
+import math
 import subprocess
 import sys
+import time
 
 import pytest
 from datasets import Dataset as HFDataset
@@ -229,3 +232,245 @@ def test_constructor_rejects_too_many_examples():
 def test_constructor_rejects_non_integer_k(n_shot):
     with pytest.raises(TypeError, match="n_shot must be an int"):
         _task(n_shot=n_shot)
+
+
+# ---------------------------------------------------------------------------
+# safe_eval — two properties that pull against each other, which is why both
+# are here: it must not execute model output, and it must still agree with the
+# bare `eval` upstream uses on every expression a real answer produces.
+# ---------------------------------------------------------------------------
+
+#: Upstream's namespace: number_utils.py's module globals *plus* the real
+#: builtins, which is what a bare `eval(num)` sees. Omitting "__builtins__" is
+#: what restores the latter — Python injects them.
+_UPSTREAM_GLOBALS = {
+    "math": math,
+    "sqrt": math.sqrt,
+    "sin": math.sin,
+    "cos": math.cos,
+    "log": math.log,
+    "pi": math.pi,
+    "factorial": math.factorial,
+    "exp": math.exp,
+    "e": math.e,
+    "E": 2.718,
+}
+
+#: The payload from the bug report: a walk from a literal back to a live
+#: `open` that a cleared `__builtins__` does not stop.
+_ESCAPE = (
+    "[c for c in ().__class__.__base__.__subclasses__() "
+    "if c.__name__=='catch_warnings'][0]()._module.__builtins__['open']"
+    "('{path}','w')"
+)
+
+
+def _safe_eval():
+    return _task_module().safe_eval
+
+
+def _unsafe_expression():
+    return _task_module().UnsafeExpression
+
+
+def test_reported_escape_payload_is_refused(tmp_path):
+    """The reported RCE is refused, and writes nothing."""
+    target = tmp_path / "pwned"
+    with pytest.raises(_unsafe_expression()):
+        _safe_eval()(_ESCAPE.format(path=target))
+    assert not target.exists()
+
+
+@pytest.mark.parametrize(
+    "entry_point",
+    ["number_it", "answer_clean", "compare_answer_with_groundtruth"],
+)
+def test_escape_is_refused_through_every_grading_entry_point(entry_point, tmp_path):
+    """The payload is reachable from an extracted answer, not just from eval."""
+    module = _task_module()
+    target = tmp_path / entry_point
+    payload = _ESCAPE.format(path=target)
+    if entry_point == "number_it":
+        module.number_it(payload)
+    elif entry_point == "answer_clean":
+        module.answer_clean(["The answer is"], f"The answer is {payload}")
+    else:
+        module.compare_answer_with_groundtruth(f"({payload},)", "[1]", [1])
+    assert not target.exists()
+
+
+def test_cleared_builtins_eval_really_was_escapable(tmp_path):
+    """The guard this replaced was not a sandbox — the reason for the change.
+
+    Pins the premise rather than the fix: if this ever stops escaping, the
+    justification for diverging from upstream has changed and should be re-read.
+    """
+    target = tmp_path / "pwned"
+    builtins.eval(_ESCAPE.format(path=target), {"__builtins__": {}})  # noqa: S307
+    assert target.exists()
+
+
+@pytest.mark.parametrize(
+    "expression",
+    [
+        "().__class__",
+        "().__class__.__base__.__subclasses__()",
+        "[c for c in (1, 2)][0]",
+        "{c for c in (1, 2)}",
+        "(c for c in (1, 2))",
+        "[1, 2][0]",
+        "lambda: 1",
+        "__import__('os')",
+        "open('x', 'w')",
+        "'a string'",
+        "f'{1}'",
+        "1 if True else 2",
+        "1 < 2",
+        "True and False",
+        "*[1],",
+        "{'a': 1}",
+        "math.__dict__",
+        "math._x",
+        "(x := 1)",
+    ],
+)
+def test_execution_shapes_are_refused(expression):
+    """Nothing that could reach an object or a statement is evaluable."""
+    with pytest.raises((_unsafe_expression(), SyntaxError)):
+        _safe_eval()(expression)
+
+
+def test_unsafe_expression_is_a_value_error():
+    """Call sites catch broad `Exception`; this keeps the eval-era contract.
+
+    A refusal has to land in the same `except` that a `SyntaxError` or a
+    `NameError` from `eval` landed in, so the answer simply fails to become a
+    number instead of failing the sample.
+    """
+    assert issubclass(_unsafe_expression(), ValueError)
+
+
+@pytest.mark.parametrize(
+    "expression",
+    [
+        "factorial(2000000)",
+        "9**9**9**9",
+        "10**10**10",
+        "(10**999)**999",
+        "1 << 10**10",
+    ],
+)
+def test_unbounded_computation_is_refused_promptly(expression):
+    """Grading shares one event loop, so an unbounded answer stalls a session."""
+    started = time.perf_counter()
+    with pytest.raises(_unsafe_expression()):
+        _safe_eval()(expression)
+    assert time.perf_counter() - started < 1.0
+
+
+def test_node_count_is_bounded():
+    module = _task_module()
+    with pytest.raises(module.UnsafeExpression, match="too large"):
+        module.safe_eval("+".join(["1"] * (module._MAX_NODES + 10)))
+
+
+def test_factorial_is_bounded_but_usable():
+    module = _task_module()
+    cap = module._MAX_FACTORIAL
+    assert module.safe_eval(f"factorial({cap})") == math.factorial(cap)
+    with pytest.raises(module.UnsafeExpression):
+        module.safe_eval(f"factorial({cap + 1})")
+
+
+#: Expressions drawn from the shapes the three call sites actually reach:
+#: latex2sympy's `str()` output, and answers a model writes directly.
+_FAITHFUL = [
+    "1",
+    "-1",
+    "1.5",
+    "-3.25e-7",
+    "1 + 2",
+    "3 - 4",
+    "2*3",
+    "7/2",
+    "7//2",
+    "7%2",
+    "2**10",
+    "-2**3",
+    "+5",
+    "(1 + 2)*3",
+    "[1, 2, 3]",
+    "(1, 2, 3)",
+    "[0, 0]",
+    "{0}",
+    "{1, 2}",
+    "sqrt(2)",
+    "sin(0)",
+    "cos(0)",
+    "log(1)",
+    "exp(1)",
+    "pi",
+    "e",
+    "E",
+    "factorial(5)",
+    "math.sqrt(9)",
+    "math.pi",
+    "3.54*E - 7",
+    "E/(-1 + E)",
+    # builtins upstream exposed; the cleared-namespace eval lost every one
+    "abs(-5)",
+    "round(1.6)",
+    "pow(2, 10)",
+    "min(3, 1)",
+    "max(3, 1)",
+    "sum([1, 2, 3])",
+    "int(2.9)",
+    "float(3)",
+    "divmod(7, 2)",
+    "len([1, 2])",
+    # bitwise: kept purely because upstream computes them
+    "5 ^ 3",
+    "5 & 3",
+    "5 | 3",
+    "1 << 4",
+    "16 >> 2",
+    "~5",
+]
+
+
+@pytest.mark.parametrize("expression", _FAITHFUL)
+def test_matches_upstream_bare_eval(expression):
+    """Same value as upstream's `eval(num)` for every evaluable shape."""
+    expected = builtins.eval(expression, dict(_UPSTREAM_GLOBALS))  # noqa: S307
+    assert _safe_eval()(expression) == expected
+
+
+@pytest.mark.parametrize(
+    "expression",
+    ["a", "x_2", "oo", "Interval", "Integral", "unknown_name"],
+)
+def test_unknown_names_fail_as_upstream_does(expression):
+    """Symbolic leftovers from latex2sympy: upstream raises NameError, so do we.
+
+    The outcome, not the exception type, is the contract — both readings leave
+    the answer un-numeric, which is what the call sites branch on.
+    """
+    with pytest.raises(_unsafe_expression(), match="unknown name"):
+        _safe_eval()(expression)
+    with pytest.raises(NameError):
+        builtins.eval(expression, dict(_UPSTREAM_GLOBALS))  # noqa: S307
+
+
+def test_set_display_is_evaluated_not_refused():
+    """Fidelity where safety does not object — the set-display defect stands.
+
+    Refusing `{0}` would rescue a list answer latex2sympy folded into a set and
+    score +2/800 on the measured run. That is a grader repair, so it belongs to
+    a `_fixed` variant and deliberately does not happen under this name.
+    """
+    assert _safe_eval()("{0}") == {0}
+
+
+def test_syntax_error_propagates_as_from_eval():
+    with pytest.raises(SyntaxError):
+        _safe_eval()("1 +")
